moved task to a separate file. fixed a small bug
