+++ src/headphones2/postprocess/__init__.py
@@ -2,3 +2,5 @@ from headphones2.postprocess.component_base import PostProcessorException
 from headphones2.postprocess.extensions.renamer import Renamer
 from headphones2.postprocess.taggers.acoustid_tagger import AcoustIDAlbumTagger
 from headphones2.postprocess.taggers.simple_tagger import BeetsTagger
+
+from headphones2.postprocess.task import AlbumTask
