+++ src/headphones2/postprocess/taggers/simple_tagger.py
@@ -1,6 +1,5 @@
 from __future__ import (absolute_import, division,
                         print_function, unicode_literals)
-from collections import namedtuple
 
 import logbook
 
@@ -8,9 +7,6 @@ from beets.autotag import tag_album, Recommendation
 from headphones2.postprocess.component_base import PostProcessor, PostProcessorException
 
 logger = logbook.Logger(__name__)
-
-BeetsTaggerResult = namedtuple("Result",
-                               ["album_id", "album_info_object", "track_mapping_object", "extra_items", "extra_tracks"])
 
 
 class BeetsTagger(PostProcessor):
@@ -45,6 +41,9 @@ class BeetsTagger(PostProcessor):
             for tagger in fallback_taggers:
                 logger.debug("Calling {}".format(tagger.__class__.__name__))
                 fallback_recommendation = tagger.process(task)
+                if not fallback_recommendation:
+                    continue
+
                 artist_name, album_name, album_recommendation_list, recommendation = \
                     tag_album(task.items, search_artist=expected_artist, search_album=expected_album,
                               search_id=fallback_recommendation)
@@ -62,6 +61,7 @@ class BeetsTagger(PostProcessor):
         task._track_mapping_object = track_mapping
 
         return True
+
 
 class BeetsTaggerException(PostProcessorException):
     pass
